initialIngest
The Catalog Services for the Web Recommendation: ISO 19115 Profile has
three different parts. I have chosen to make each of these parts a
suite. In this xml they are wrapped in a root element. Some of these
checks are required and some are optional. This would be akin to taking
the LTER recommendation and presenting the identification level, etc as
separate suites. This allows the user to test for only the part of the
recommendation they are interested in.

Would it be simpler if each of these related suites was a separate file?