Add 'k8s.cluster` value to all metadig-* values
The `--set k8s.cluster="dev"` helm install option causes metadig config files to be installed that are appropriate for the k8s dev environment, as they cause harvesting on the DataONE stage env, not production. Add this value to all locally authoried metadig-* charts so it can be specified for all charts, for consistency.